Refactor: MessageTree/Renderer to twitch.tv site
